# swagger/movie-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Movies-API",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register User",
        "description": "Register User",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "susan@gmail.com"
                  },
                  "name": {
                    "type": "string",
                    "example": "susan"
                  },
                  "password": {
                    "type": "string",
                    "example": "secret"
                  }
                }
              },
              "examples": {
                "Register User": {
                  "value": {
                    "email": "susan@gmail.com",
                    "name": "susan",
                    "password": "secret"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login User",
        "description": "Login User",
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "susan@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "secret"
                  }
                }
              },
              "examples": {
                "Login User": {
                  "value": {
                    "email": "susan@gmail.com",
                    "password": "secret"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/movies": {
      "get": {
        "summary": "Get All Movies for User",
        "description": "Retrieves all movies associated with the currently authenticated user. Requires a valid Bearer token obtained from the /auth/login endpoint.",
        "operationId": "getAllMovies",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create Movie",
        "description": "Create Movie",
        "operationId": "createMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Thunderbolts*"
                  },
                  "director": {
                    "type": "string",
                    "example": "Jake Schreier"
                  },
                  "year": {
                    "type": "number",
                    "example": 2025
                  },
                  "plot": {
                    "type": "string",
                    "example": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts."
                  }
                }
              },
              "examples": {
                "Create Movie": {
                  "value": {
                    "title": "Thunderbolts*",
                    "director": "Jake Schreier",
                    "year": 2025,
                    "plot": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/movies/{movieId}": {
      "parameters": [
        {
          "name": "movieId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of the movie to fetch, update, or delete"
        }
      ],
      "get": {
        "summary": "Get Single Movie for User",
        "description": "Retrieves a single movie by ID associated with the currently authenticated user. Requires a valid Bearer token.",
        "operationId": "getSingleMovie",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update Movie for User",
        "description": "Updates a movie by ID associated with the currently authenticated user. Requires a valid Bearer token.",
        "operationId": "updateMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                   "title": {
                    "type": "string",
                    "example": "Thunderbolts*"
                  },
                  "director": {
                    "type": "string",
                    "example": "Jack Schreier"
                  },
                  "year": {
                    "type": "number",
                    "example": 2025
                  },
                  "plot": {
                    "type": "string",
                    "example": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts."
                  }
                }
              },
              "examples": {
                "Update Movie": {
                  "value": {
                    "title": "Thunderbolts*",
                    "director": "Jack Schreier",
                    "year": 2025,
                    "plot": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Movie for User",
        "description": "Deletes a movie by ID associated with the currently authenticated user. Requires a valid Bearer token.",
        "operationId": "deleteMovie",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": []
}